アイテムの使用場所で分割
