+++ app/src/main/java/com/sbkinoko/sbkinokorpg/battleframe/window/window_chose_target/BattleWindow_TOP.java
@@ -11,8 +11,8 @@ import androidx.core.content.res.ResourcesCompat;
 import com.sbkinoko.sbkinokorpg.GameParams;
 import com.sbkinoko.sbkinokorpg.MainGame;
 import com.sbkinoko.sbkinokorpg.R;
-import com.sbkinoko.sbkinokorpg.game_item.action_item.UseItem;
 import com.sbkinoko.sbkinokorpg.battleframe.BattleSystem;
+import com.sbkinoko.sbkinokorpg.game_item.action_item.use_item.UseItem;
 
 public class BattleWindow_TOP extends Window_ChooseTarget {
     boolean allSelectFlag = false;
