+++ app/src/main/java/com/sbkinoko/sbkinokorpg/mapframe/event/MapEvent.java
@@ -11,7 +11,7 @@ import com.sbkinoko.sbkinokorpg.GameParams;
 import com.sbkinoko.sbkinokorpg.OptionConst;
 import com.sbkinoko.sbkinokorpg.battleframe.status.PlayerStatus;
 import com.sbkinoko.sbkinokorpg.dataList.item.List_Tool;
-import com.sbkinoko.sbkinokorpg.game_item.action_item.UseItem;
+import com.sbkinoko.sbkinokorpg.game_item.action_item.use_item.UseItemInField;
 import com.sbkinoko.sbkinokorpg.mapframe.MapBackgroundCell;
 import com.sbkinoko.sbkinokorpg.mapframe.MapFrame;
 import com.sbkinoko.sbkinokorpg.mapframe.Player;
@@ -171,13 +171,13 @@ public class MapEvent {
         roadMap(townID);
 
         mapFrame.getMapTextBoxWindow().openMenu(new String[]{"ワープした"});
-        UseItem.useInField(
+        UseItemInField.useInField(
                 groupOfWindows,
                 null,
                 this);
     }
 
-    public void warp() {
+    public void openWarpMenu() {
         String name, itemName;
         PlayerStatus _fromPlayer = mapFrame.getWindowGroupInfo().getFromPlayerStatus();
         if (_fromPlayer == null) {
