+++ app/src/main/java/com/sbkinoko/sbkinokorpg/mapframe/window/window_set/list_strategy/use_status/StrategyForUseItem.java
@@ -1,7 +1,7 @@
 package com.sbkinoko.sbkinokorpg.mapframe.window.window_set.list_strategy.use_status;
 
 import com.sbkinoko.sbkinokorpg.battleframe.status.PlayerStatus;
-import com.sbkinoko.sbkinokorpg.game_item.action_item.UseItem;
+import com.sbkinoko.sbkinokorpg.game_item.action_item.use_item.UseItem;
 
 public class StrategyForUseItem extends StrategyForNeedStatus {
